Oh my god. Citation schemas are more complex
This was intended as a simple script to be able to get citations
working in a simple manner, just to find out, that the citations
aren't really just concatenations of simple strings but concatenations
of groups of strings that might be missing totally, or exist partially.

The docstring contains lots of examples, I think it can be implemented
like this. This is for #2174

git-svn-id: https://svn.syslab.com/svn/syslabcom/recensio.policy/trunk@6862 51ee9ba4-e84d-4896-84de-354f1821c169